+++ gsad/src/gsad.c
@@ -304,7 +304,6 @@ init_validator ()
                      "|(download_ca_pub)"
                      "|(download_key_pub)"
                      "|(edit_alert)"
-                     "|(edit_config)"
                      "|(edit_config_family)"
                      "|(edit_config_nvt)"
                      "|(edit_role)"
@@ -373,6 +372,7 @@ init_validator ()
                      "|(get_info)"
                      "|(get_note)"
                      "|(get_notes)"
+                     "|(get_nvt_families)"
                      "|(get_override)"
                      "|(get_overrides)"
                      "|(get_permission)"
@@ -1918,7 +1918,6 @@ exec_gmp_get (http_connection_t *con, gsad_connection_info_t *con_info,
   ELSE (get_task)
   ELSE (get_tasks)
   ELSE (edit_alert)
-  ELSE (edit_config)
   ELSE (edit_config_family)
   ELSE (edit_config_nvt)
   ELSE (edit_role)
@@ -2064,6 +2063,7 @@ exec_gmp_get (http_connection_t *con, gsad_connection_info_t *con_info,
   ELSE (get_info)
   ELSE (get_note)
   ELSE (get_notes)
+  ELSE (get_nvt_families)
   ELSE (get_override)
   ELSE (get_overrides)
   ELSE (get_permission)