+++ gsad/src/gsad_gmp.h
@@ -314,9 +314,6 @@ save_target_gmp (gvm_connection_t *, credentials_t *, params_t *,
                  cmd_response_data_t *);
 
 char *
-edit_config_gmp (gvm_connection_t *, credentials_t *, params_t *,
-                 cmd_response_data_t *);
-char *
 get_config_gmp (gvm_connection_t *, credentials_t *, params_t *,
                 cmd_response_data_t *);
 char *
@@ -412,6 +409,10 @@ export_note_gmp (gvm_connection_t *, credentials_t *, params_t *,
 char *
 export_notes_gmp (gvm_connection_t *, credentials_t *, params_t *,
                   cmd_response_data_t *);
+
+char *
+get_nvt_families_gmp (gvm_connection_t *, credentials_t *, params_t *,
+                      cmd_response_data_t *);
 
 char *
 create_permission_gmp (gvm_connection_t *, credentials_t *, params_t *,